requirements.txt
selenium==4.10.0
webdriver-manager==3.8.6
requests==2.31.0
schedule==1.2.0
argparse==1.4.0
python-dotenv==1.0.0

